Remove test for error code returned when sending empty message
This relies on some validation code [1] in smtp-connection
to check if the `message` parameter passed to it is not falsy.
In practice this never happens as the message passed to it is always
a stream as create by MimeNode.prototype.createReadStream [2]
The introduction of a stream-based MockBuilder has revealed this test
as incorrect.

[1] https://github.com/andris9/smtp-connection/blob/914b20fd/src/smtp-connection.js#L269-L271
[2] https://github.com/andris9/nodemailer-smtp-transport/blob/70e53c9a/src/smtp-transport.js#L67